feat: extract link snippet utility
